Move markdown serialization off the shared mention extension
The sentinel is a storage concern of the fieldtype, not a property of the
mention node, and its output is only valid alongside the escaping pass in
markdown.js. Declaring it there keeps the extension reusable and leaves
the primitive's imports free of markdown.
